spy player

+++ back/database/migrations/1700861155830_create_maps_table.ts
@@ -23,7 +23,7 @@ export default class extends BaseSchema {
             table.decimal('finance_wild_territory_cost_factor').notNullable();
             table.decimal('finance_wild_territory_enforcement_factor').notNullable();
             table.decimal('wild_territory_subversion_factor').notNullable();
-            table.decimal('base_subversion_cost').notNullable();
+            table.integer('base_subversion_cost').notNullable();
             table.decimal('subversion_cost_factor').notNullable();
             table.integer('fortify_cost').notNullable();
             table.uuid('neutral_flag_id').notNullable().references('id').inTable('files').onDelete('CASCADE');
